+++ src/main/java/de/kamillionlabs/hateoflux/linkbuilder/BaseUrlExtractor.java
@@ -1,0 +1,66 @@
+/*
+ * Copyright (c)  2024 kamillion-suite contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * @since 17.10.2024
+ */
+
+package de.kamillionlabs.hateoflux.linkbuilder;
+
+import org.springframework.http.server.reactive.ServerHttpRequest;
+import org.springframework.lang.NonNull;
+import org.springframework.web.server.ServerWebExchange;
+
+import java.net.URI;
+
+/**
+ * Utility class for extracting the base URL from a server web exchange or HTTP request.
+ *
+ * @author Younes El Ouarti
+ */
+public class BaseUrlExtractor {
+
+    /**
+     * Extracts the base URL from the given {@link ServerWebExchange}.
+     *
+     * @param exchange
+     *         the server web exchange
+     * @return the base URL as a string
+     */
+    public static String extractBaseUrl(@NonNull ServerWebExchange exchange) {
+        return extractBaseUrl(exchange.getRequest());
+    }
+
+    /**
+     * Extracts the base URL from the given {@link ServerHttpRequest}.
+     *
+     * @param request
+     *         the server HTTP request
+     * @return the base URL as a string
+     */
+    public static String extractBaseUrl(@NonNull ServerHttpRequest request) {
+        URI uri = request.getURI();
+
+        String scheme = uri.getScheme();
+        String host = uri.getHost();
+        int port = uri.getPort();
+
+        boolean includePort = (port != -1 && ((scheme.equals("http") && port != 80)
+                || (scheme.equals("https") && port != 443)));
+        String portPart = includePort ? ":" + port : "";
+
+        return scheme + "://" + host + portPart;
+    }
+}
+