+++ src/test/java/de/kamillionlabs/hateoflux/assembler/ReactiveFlatHalWrapperAssemblerTest.java
@@ -22,12 +22,12 @@ class ReactiveFlatHalWrapperAssemblerTest {
 
         @Override
         public Link buildSelfLinkForEntityList(ServerWebExchange exchange) {
-            return Link.linkAsSelfOf("reactive/entity-list/self/link");
+            return Link.of("reactive/entity-list/self/link");
         }
 
         @Override
         public Link buildSelfLinkForEntity(Book entityToWrap, ServerWebExchange exchange) {
-            return Link.linkAsSelfOf("reactive/entity/self/link");
+            return Link.of("reactive/entity/self/link");
         }
     }